refactor(types): refactor Block content types
Rename content types from generic Block* to PortableText* as that's what it really is.
Refactor the types to better type out PortableText in general and as spesific as possible
according to the PT spec. A good example would be a PortableTextBlock which can be
either a PortableTextTextBlock or a PortableTextObject. The PT text itself would be
a PortableTextTextBlock, while an image or some other object in the PT-array would
be a PortableTextObject.

This also helps with exports to ts-doc etc, so that the PT-types are grouped under
each other.

These types will be the single source of truth for PortableText elsewhere in the Studio.